CHORE Front moved agenda from secretary components to shared components
