+++ client/src/components/shared_components/Agenda/events.utils.ts
